+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2025-01-27
+
 ### Added
 
-- Initial release
+- Added a document command to update the list of used and available linters
+- Added Swiftlint build phase to Xcode projects
+- Added MegaLinter for linting common file formats with generator command and rake task
+- Added a GitHub audit rule to check that linting is used and configured correctly
+
+[unreleased]: https://github.com/HealthDataInsight/way_of_working-code_linting-hdi/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/HealthDataInsight/way_of_working-code_linting-hdi/releases/tag/v1.0.0